requirements.txt
PySide6-Essentials==6.9.0
pyqtgraph==0.13.7
rich==14.0.0